+++ libc/uclibc-ng/libc/ARCH-all/libc/stdlib/random_sys.c
@@ -2,7 +2,7 @@
 #include <sys/random.h>
 #include <errno.h>
 
-int getrandom(void *__buf, size_t count, unsigned int flags)
+ssize_t getrandom(void *__buf, size_t count, unsigned int flags)
 {
   /* Could implement using L4Re::random here */
   return -ENODEV;
